구매

+++ src/main/java/com/dh/home/product/ProductImageVO.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 public class ProductImageVO {
 	
-	private int fileNum;
+	private Long fileNum;
 	private String fileName;
 	private String oriName;
-	private int itemNum;
+	private Long itemNum;
 }
